Debounce the input to capture only 'real' button interactions.
Due to the physical and mechanical nature of the button it can generate
bogus state transitions.  We can compensate for this by checking twice
in a short period of time to make sure we are dealing with a real button
state change.  50ms seems about right to not have false read, but to
_do_ register quicker consecutive button presses, which we'll need if we
ever want to translate Morse Code ;-)